[fix] errorfix

+++ src/main/java/com/campusform/server/recruiting/presentation/CommentController.java
@@ -5,9 +5,12 @@ import com.campusform.server.recruiting.application.service.CommentService;
 import com.campusform.server.recruiting.application.dto.request.CommentRequest;
 import com.campusform.server.recruiting.application.dto.response.CommentCreateResponse;
 import com.campusform.server.recruiting.application.dto.response.CommentUpdateResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
+//import org.springframework.security.core.Authentication;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -20,13 +23,16 @@ public class CommentController {
     // 댓글 작성
     @PostMapping
     public ResponseEntity<CommentCreateResponse> createComment(
-        @PathVariable Long projectId,
+        //@PathVariable Long projectId,
         @PathVariable Long applicantId,
-        @RequestParam(defaultValue = "document") String stage,
-        @RequestBody CommentRequest request,
-        Authentication authentication
+        //@RequestParam StageStatus stage,
+        @RequestBody @Valid CommentRequest request,
+        @AuthenticationPrincipal OAuth2User oauth2User
     ){
-        Long memberId = authService.extractUserId(authentication);
+        if (oauth2User == null) {
+            throw new IllegalArgumentException("로그인이 필요합니다.");
+        }
+        Long memberId = oauth2User.getAttribute("memberId");
         CommentCreateResponse response = commentService.createComment(applicantId, memberId, request);
         return ResponseEntity.ok(response);
     }
@@ -34,17 +40,20 @@ public class CommentController {
     // 댓글 수정
     @PatchMapping("/{commentId}")
     public ResponseEntity<CommentUpdateResponse> updateComment(
-            @PathVariable Long projectId,
+            //@PathVariable Long projectId,
             @PathVariable Long applicantId,
             @PathVariable Long commentId,
-            @RequestParam(defaultValue = "document") String stage,
-            @RequestBody CommentRequest request,
-            Authentication authentication
+            //@RequestParam StageStatus stage,
+            @RequestBody @Valid CommentRequest request,
+            @AuthenticationPrincipal OAuth2User oauth2User
     ) {
-        Long memberId = authService.extractUserId(authentication);
+        if (oauth2User == null) {
+            throw new IllegalArgumentException("로그인이 필요합니다.");
+        }
+        Long memberId = oauth2User.getAttribute("userId");
 
         CommentUpdateResponse response = commentService.updateComment(
-                projectId, applicantId,commentId, memberId, request
+                applicantId,commentId, memberId, request
         );
 
         return ResponseEntity.ok(response);
@@ -53,14 +62,16 @@ public class CommentController {
     // 댓글 삭제
     @DeleteMapping("/{commentId}")
     public ResponseEntity<?> deleteComment(
-            @PathVariable Long projectId,
-            @PathVariable Long applicantId,
+            // @PathVariable Long projectId,
+            // @PathVariable Long applicantId,
             @PathVariable Long commentId,
-            @RequestParam(defaultValue = "document") String stage,
-            Authentication authentication
+            // @RequestParam StageStatus stage,
+            @AuthenticationPrincipal OAuth2User oauth2User
     ) {
-        Long memberId = authService.extractUserId(authentication);
-
+        if (oauth2User == null) {
+            throw new IllegalArgumentException("로그인이 필요합니다.");
+        }
+        Long memberId = oauth2User.getAttribute("userId");
         commentService.deleteComment(commentId, memberId);
 
         return ResponseEntity.ok().build();
